feat: enhance HomeScreen UI with sticky header and transaction grouping
This commit introduces several UI enhancements to the HomeScreen:

-   A sticky header is implemented, combining the account summary and toolbar.
-   Transactions are now grouped by date ("Today", "Yesterday", or specific date).
-   The overall visual design of the HomeScreen has been refined.
-   Updated `AccountSummaryUiModel` to use `BigDecimal` for balance.
-   The default transaction name now uses the category name if no specific name is provided.
-   The transaction item now displays `toAccountName`, `merchantName`, or `fromAccountName` before falling back to `name`.

+++ app/src/main/java/dev/pandesal/sbp/presentation/home/HomeViewModel.kt
@@ -74,7 +74,7 @@ private fun dev.pandesal.sbp.domain.model.Account.toUiModel(): AccountSummaryUiM
     val isFundingWallet = type == dev.pandesal.sbp.domain.model.AccountType.BANK_ACCOUNT
     return AccountSummaryUiModel(
         name = name,
-        balance = balance.toDouble(),
+        balance = balance,
         isSpendingWallet = isSpendingWallet,
         isFundingWallet = isFundingWallet,
         currency = currency
